macOS: Silence warning about supporting secure state restoration
We don't support or implement state restorations via the AppKit
state restoration APIs, but if we did, we would/should support
secure state restoration. This is the default for apps linked
against the macOS 14 SDK, but as we target versions below that
as well we need to return YES here explicitly to silence a runtime
warning.

Pick-to: 6.5 6.2
Fixes: QTBUG-117745
Change-Id: I0145504a79e53499852832d23dc7d4d6838dfa1b
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit fc1549c01445bb9c99d3ba6de8fa9da230614e72)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>