Android: add @UsedFromNativeCode annotation for java method used in c++
To get rid of unused method warnings in Java code for methods used only
from C++ code and make it more visible which methods are used by native
code.

This annotation would need to be added to the relevant methods in this
patch or future one.

Use this annotation for all methods called from native code. And along
the way remove few methods that are not used anywhere.

Task-number: QTBUG-118077
Change-Id: Ie493fb057f4a93aff513f85e0936bc65a9bb304c
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>